implement 1/f noise removal
