+++ spaceKLIP/imagetools.py
@@ -2105,149 +2105,3 @@ def cube_outlier_detection(data, sigma_cut=5, nint_min=5):
     indbad = ~indgood
 
     return indbad
-
-
-def cube_fit(tarr, data, sat_vals, sat_frac=0.95, bias=None, 
-             deg=1, bpmask_arr=None, fit_zero=False, verbose=False,
-             use_legendre=False, lxmap=None, return_lxmap=False,
-             return_chired=False):
-    """Fit unsaturated data and return coefficients"""
-        
-    from webbpsf_ext.maths import jl_poly_fit, jl_poly
-
-    nz, ny, nx = data.shape
-    
-    # Subtract bias?
-    imarr = data if bias is None else data - bias
-        
-    # Array of masked pixels (saturated)
-    mask_good = imarr < sat_frac*sat_vals
-    if bpmask_arr is not None:
-        mask_good = mask_good & ~bpmask_arr
-    
-    # Reshape for all pixels in single dimension
-    imarr = imarr.reshape([nz, -1])
-    mask_good = mask_good.reshape([nz, -1])
-
-    # Initial 
-    cf = np.zeros([deg+1, nx*ny])
-    if return_lxmap:
-        lx_min = np.zeros([nx*ny])
-        lx_max = np.zeros([nx*ny])
-    if return_chired:
-        chired = np.zeros([nx*ny])
-
-    # For each 
-    npix_sum = 0
-    i0 = 0 if fit_zero else 1
-    for i in np.arange(i0,nz)[::-1]:
-        ind = (cf[1] == 0) & (mask_good[i])
-        npix = np.sum(ind)
-        npix_sum += npix
-        
-        if verbose:
-            print(i+1,npix,npix_sum, 'Remaining: {}'.format(nx*ny-npix_sum))
-            
-        if npix>0:
-            if fit_zero:
-                x = np.concatenate(([0], tarr[0:i+1]))
-                y = np.concatenate((np.zeros([1, np.sum(ind)]), imarr[0:i+1,ind]), axis=0)
-            else:
-                x, y = (tarr[0:i+1], imarr[0:i+1,ind])
-
-            if return_lxmap:
-                lx_min[ind] = np.min(x) if lxmap is None else lxmap[0]
-                lx_max[ind] = np.max(x) if lxmap is None else lxmap[1]
-                
-            # Fit line if too few points relative to polynomial degree
-            if len(x) <= deg+1:
-                cf[0:2,ind] = jl_poly_fit(x,y, deg=1, use_legendre=use_legendre, lxmap=lxmap)
-            else:
-                cf[:,ind] = jl_poly_fit(x,y, deg=deg, use_legendre=use_legendre, lxmap=lxmap)
-
-            # Get reduced chi-sqr metric for poorly fit data
-            if return_chired:
-                yfit = jl_poly(x, cf[:,ind])
-                deg_chi = 1 if len(x)<=deg+1 else deg
-                dof = y.shape[0] - deg_chi
-                chired[ind] = chisqr_red(y, yfit=yfit, dof=dof)
-
-    imarr = imarr.reshape([nz,ny,nx])
-    mask_good = mask_good.reshape([nz,ny,nx])
-    
-    cf = cf.reshape([deg+1,ny,nx])
-    if return_lxmap:
-        lxmap_arr = np.array([lx_min, lx_max]).reshape([2,ny,nx])
-        if return_chired:
-            chired = chired.reshape([ny,nx])
-            return cf, lxmap_arr, chired
-        else:
-            return cf, lxmap_arr
-    else:
-        if return_chired:
-            chired = chired.reshape([ny,nx])
-            return cf, chired
-        else:
-            return cf
-        
-def chisqr_red(yvals, yfit=None, err=None, dof=None,
-               err_func=np.std):
-    """ Calculate reduced chi square metric
-    
-    If yfit is None, then yvals assumed to be residuals.
-    In this case, `err` should be specified.
-    
-    Parameters
-    ==========
-    yvals : ndarray
-        Sampled values.
-    yfit : ndarray
-        Model fit corresponding to `yvals`.
-    dof : int
-        Number of degrees of freedom (nvals - nparams - 1).
-    err : ndarray or float
-        Uncertainties associated with `yvals`. If not specified,
-        then use yvals point-to-point differences to estimate
-        a single value for the uncertainty.
-    err_func : func
-        Error function uses to estimate `err`.
-    """
-    
-    if (yfit is None) and (err is None):
-        print("Both yfit and err cannot be set to None.")
-        return
-    
-    diff = yvals if yfit is None else yvals - yfit
-    
-    sh_orig = diff.shape
-    ndim = len(sh_orig)
-    if ndim==1:
-        if err is None:
-            err = err_func(yvals[1:] - yvals[0:-1]) / np.sqrt(2)
-        dev = diff / err
-        chi_tot = np.sum(dev**2)
-        dof = len(chi_tot) if dof is None else dof
-        chi_red = chi_tot / dof
-        return chi_red
-    
-    # Convert to 2D array
-    if ndim==3:
-        sh_new = [sh_orig[0], -1]
-        diff = diff.reshape(sh_new)
-        yvals = yvals.reshape(sh_new)
-        
-    # Calculate errors for each element
-    if err is None:
-        err_arr = np.array([yvals[i+1] - yvals[i] for i in range(sh_orig[0]-1)])
-        err = err_func(err_arr, axis=0) / np.sqrt(2)
-        del err_arr
-    else:
-        err = err.reshape(diff.shape)
-    # Get reduced chi sqr for each element
-    dof = sh_orig[0] if dof is None else dof
-    chi_red = np.sum((diff / err)**2, axis=0) / dof
-    
-    if ndim==3:
-        chi_red = chi_red.reshape(sh_orig[-2:])
-        
-    return chi_red
